Move contents of hmac-sha2-256 and hmac-sha2-512 modules to hmac-sha2
