+++ gradle.properties
@@ -29,10 +29,10 @@ POM_DEVELOPER_ID=KotlinCrypto
 POM_DEVELOPER_NAME=Kotlin Crypto
 POM_DEVELOPER_URL=https://github.com/KotlinCrypto/
 
-VERSION_NAME=0.1.2-SNAPSHOT
+VERSION_NAME=0.2.0-SNAPSHOT
 # 0.1.0-alpha01 = 00 01 00 11
 # 0.1.0-beta01  = 00 01 00 21
 # 0.1.0-rc01    = 00 01 00 31
 # 0.1.0         = 00 01 00 99
 # 1.1.0         = 01 01 00 99
-VERSION_CODE=10299
+VERSION_CODE=20099
