--- library/hmac/hmac-sha2-256/api/hmac-sha2-256.api
@@ -1,4 +1,0 @@
-public final class org/kotlincrypto/macs/HmacSHA256 : org/kotlincrypto/macs/Hmac {
-	public fun <init> ([B)V
-}
-
