--- library/hmac/hmac-sha2-256/src/jvmTest/kotlin/org/kotlincrypto/macs/HmacSHA256JvmUnitTest.kt
@@ -1,58 +1,0 @@
-/*
- * Copyright (c) 2023 Matthew Nelson
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- **/
-package org.kotlincrypto.macs
-
-import org.kotlincrypto.core.Mac
-import kotlin.test.Test
-
-class HmacSHA256JvmUnitTest: HmacSHA256UnitTest() {
-    override fun mac(key: ByteArray): Mac = TestJvmMac(super.mac(key).algorithm(), key)
-
-    @Test
-    override fun givenMac_whenResetSmallKey_thenDoFinalReturnsExpected() {
-        super.givenMac_whenResetSmallKey_thenDoFinalReturnsExpected()
-    }
-
-    @Test
-    override fun givenMac_whenResetMediumKey_thenDoFinalReturnsExpected() {
-        super.givenMac_whenResetMediumKey_thenDoFinalReturnsExpected()
-    }
-
-    @Test
-    override fun givenMac_whenResetLargeKey_thenDoFinalReturnsExpected() {
-        super.givenMac_whenResetLargeKey_thenDoFinalReturnsExpected()
-    }
-
-    @Test
-    override fun givenMac_whenUpdatedSmall_thenDoFinalReturnsExpected() {
-        super.givenMac_whenUpdatedSmall_thenDoFinalReturnsExpected()
-    }
-
-    @Test
-    override fun givenMac_whenUpdatedMedium_thenDoFinalReturnsExpected() {
-        super.givenMac_whenUpdatedMedium_thenDoFinalReturnsExpected()
-    }
-
-    @Test
-    override fun givenMac_whenCopied_thenIsDifferentInstance() {
-        super.givenMac_whenCopied_thenIsDifferentInstance()
-    }
-
-    @Test
-    override fun givenMac_whenDoFinal_thenLengthMatchesOutput() {
-        super.givenMac_whenDoFinal_thenLengthMatchesOutput()
-    }
-}
