+++ library/hmac/hmac-sha2/api/hmac-sha2.api
@@ -1,0 +1,7 @@
+public final class org/kotlincrypto/macs/HmacSHA256 : org/kotlincrypto/macs/Hmac {
+	public fun <init> ([B)V
+}
+
+public final class org/kotlincrypto/macs/HmacSHA512 : org/kotlincrypto/macs/Hmac {
+	public fun <init> ([B)V
+}
