+++ library/hmac/hmac-sha2/gradle.properties
@@ -1,0 +1,16 @@
+# Copyright (c) 2023 Matthew Nelson
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+POM_ARTIFACT_ID=hmac-sha2
+POM_NAME=KotlinCrypto HmacSHA224, HmacSHA256, HmacSHA384, HmacSHA512, HmacSHA512/t,
+POM_DESCRIPTION=HmacSHA224, HmacSHA256, HmacSHA384, HmacSHA512, and HmacSHA512/t Message Authentication
