+++ library/hmac/hmac-sha2/hmac-sha2-256/README.md
@@ -1,0 +1,4 @@
+# hmac-sha2-256
+
+Old module contents were combined into `hmac-sha2`. This publication is solely to mitigate breaking
+others' projects and inform them they should update their dependency.
