+++ library/hmac/hmac-sha2/hmac-sha2-256/gradle.properties
