+++ library/hmac/hmac-sha2/hmac-sha2-512/gradle.properties
